// frontend/src/types/type.ts
export interface IReport {
    _id: string
    user_report: Userreport
    link: string
    type_of_violation: string
    content: string
    status: string
    is_violated: boolean
    resolved_by: string
    resolved_date: Date
    resolved_content: string
    created_at: Date
    __v: number
}

export interface IWordCount {
    learned: number
    remembered: number
    total: number
    reviewing: number
}

interface Userreport {
    _id: string
    displayName: string
    profilePicture: string
}